Resolve Issue #20 - Add nanliu-staging support
Replace mkrakowitzer-deploy with nanliu-staging as the default tar file deployment module
